Add crashylitics framework, add pdf reader, add offline reading support
Now accepts PDF/HTML files larger than 10MB.  Instead of saving file on
server, save link to file and still cache file locally.  Files are
re-downloaded from links and cached if user deletes app or moves to
different device.

Updated provisioning for TF to allow external push notifications.

+++ zither/Classes/EditProduct/SSIDatePicker.h
@@ -8,7 +8,7 @@
 
 @protocol SSIDatePickerDelegate
 
-- (void)picker:(int)tag didChooseDate:(NSDate *)date;
+- (void)picker:(NSInteger)tag didChooseDate:(NSDate *)date;
 
 @end
 
@@ -25,7 +25,7 @@
 + (SSIDatePicker *)viewWithTitle:(NSString *)title
                    defaultValue:(id)value
                        delegate:(id<SSIDatePickerDelegate>)delegate
-                            tag:(int)tag;
+                            tag:(NSInteger)tag;
 
 - (void)show;
 - (void)dismiss;
